database_requirements.txt
psycopg2-binary==2.9.9
uuid==1.30 